requirements.txt
streamlit
anthropic
beautifulsoup4
requests
gunicorn
python-dotenv
